+++ lib/service_actor/arguments_validator.rb
@@ -4,6 +4,7 @@ module ServiceActor::ArgumentsValidator
   module_function
 
   def validate_origin_name(name, origin:)
+    return if name.to_sym == :error
     return unless ServiceActor::Result.instance_methods.include?(name.to_sym)
 
     raise ArgumentError,