+++ spec/examples/handle_input_called_error.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+class HandleInputCalledError < Actor
+  input :error
+
+  def call; end
+end